HW3/requirements.txt
numpy
pillow
tifffile
scikit-image
torch
torchvision
matplotlib
scikit-learn
tqdm
pycocotools
